Rename heatmap component
