+++ src/page-search/HeatmapSection.style.js
