+++ src/page-search/SearchPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container, Headline } from './SearchPage.style';
 import SubredditForm from './SubredditForm';
-import Heatmap from './Heatmap';
+import HeatmapSection from './HeatmapSection';
 
 function SearchPage() {
   return (
@@ -11,7 +11,7 @@ function SearchPage() {
       </Headline>
 
       <SubredditForm />
-      <Heatmap />
+      <HeatmapSection />
     </Container>
   );
 }
